# PERF: O(n) speedup in any/all by re-enabling short-circuiting for bool case