@pep8speaks 2019-02-01
Hello @qwhelan! Thanks for submitting the PR.



 - In the file [`pandas/core/nanops.py`](https://github.com/pandas-dev/pandas/blob/3acc6b638a7b95f9bc7d5814d93d23b9801c5ea9/pandas/core/nanops.py), following are the PEP8 issues :

> [Line 368:9](https://github.com/pandas-dev/pandas/blob/3acc6b638a7b95f9bc7d5814d93d23b9801c5ea9/pandas/core/nanops.py#L368): [E129](https://duckduckgo.com/?q=pep8%20E129) visually indented line with same indent as next logical line
> [Line 405:9](https://github.com/pandas-dev/pandas/blob/3acc6b638a7b95f9bc7d5814d93d23b9801c5ea9/pandas/core/nanops.py#L405): [E129](https://duckduckgo.com/?q=pep8%20E129) visually indented line with same indent as next logical line