@jbrockmendel 2019-02-01
Does the implementation become any simpler after Panel is removed?