@qwhelan 2019-03-22
Apologies for the delay - interviewing processes have gotten in the way here.

@jbrockmendel Not especially.

@jreback Rebased on master. I've kept my additions in response to your comments as a separate commit for now - will squash if it seems they're going in the right general direction.

My PR to fix the underlying performance bug is here https://github.com/numpy/numpy/pull/12988 and pending merge; hopefully it will make it into numpy 1.17 and to a large extent obviate the need for this change. It may make sense to expand this PR to include all integer `dtypes`, as they also cannot store `NaN`.

Additionally, `np.any()` and `np.all()` attempt to short-circuit but actually exhibit `O(n)` behavior due to processing of reduction ufuncs in `NPY_BUFSIZE` (ie, 8192 byte) chunks. I'm investigating ways to remove or increase the buffer size there, which should make `pd.all()` and `pd.any()` considerably faster. It's worth noting that there's no attempt to short-circuit for any non-`bool` dtypes.